Clean up (so far) for the mobile version of World View
My edits so far on cleaning up the existing code of World View. The purpose of creating a new project was due to troubles editing the existing project. Trying to change one thing caused a majority of it to crash.

So a new one was created, and i was taking functionality from the old one and using it respectively. Cam functionality still need to be put into, but i should be done soon.