More UI Changes
Got rid of map control on the side menu and used a pop up style menu instead. Inserted UI settings on the side menu (actual functionality isn't present but layout is present).

Commented out the padding for the mMap, due to it causing issues with the position of the location button.